requirements.txt
flask>=2.3.2
mysql-connector-python>=8.0.33
sentence-transformers>=2.2.2
numpy>=1.26.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
flask-cors>=4.0.0
scikit-learn>=1.3.0
setuptools>=68.0.0
wheel>=0.41.0
